backend-ml/SignDataGenerator/requirements.txt
Pillow==8.1.0
numpy==1.19.3
jupyter==1.0.0